[Google Drive] Use generic GoogleWorkspaceAPI client from google.py
